// web/src/__generated__/schemas/CampaignFunnelDto.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

import type { CampaignNodeFunnelDto } from './CampaignNodeFunnelDto';

export type CampaignFunnelDto = { 
  campaignId?: string; 
  nodes?: CampaignNodeFunnelDto[]; 
  version?: number; 
};