// web/src/__generated__/schemas/CampaignStatusDto.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

import type { CampaignLifecycle } from './enums/CampaignLifecycle';

export type CampaignStatusDto = { 
  campaignId?: string; 
  phase?: CampaignLifecycle; 
  version?: number; 
};